Update README.md
DecSoft App Builder v2025.2

The DecSoft App Builder apps are based in HTML, CSS and jаvascript. These are the languages that you must know (more or less) in order to create modern apps that run in all the mobile and desktop browsers, as well platforms like Android, iOS and Microsoft Windows. All with the same base code! This is a big reason to choose DecSoft App Builder.

[Download DecSoft App Builder v2025.2](https://developer.team/misc-development/35244-decsoft-app-builder-v20252.html)